// src/lib/types/generated/GradedAnswer.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type GradedAnswer = { "correct": { group_id: string, marked_indices: Array<number>, } } | { "wrong": { group_id: string, marked_indices: Array<number>, correct_indices: Array<number>, } } | { "blank": { group_id: string, } } | { "multiple": { group_id: string, marked_indices: Array<number>, } } | { "partial": { group_id: string, marked_indices: Array<number>, correct_indices: Array<number>, 
/**
 * `marked_correct / total_correct`, in `(0.0, 1.0)`.
 */
score_ratio: number, } } | { "uncertain": { group_id: string, uncertain_indices: Array<number>, } };